MAGE-1203: Reorganize Builder Api folders
